requirements.txt
streamlit==1.37.0
pandas==1.5.3
pandasai==2.2.12
matplotlib==3.8.1
seaborn==0.13.1
plotly==5.24.1
Pillow==10.4.0
langchain_groq==0.1.10
langchain==0.2.1
nltk==3.9.1
wordcloud==1.9.4
python-dotenv